fix possible infinite loop problem.
In order to distinguish between random k and deterministic k, I had at a
previous point created a loop which in standard ECDSA would generate k
randomly, but here does not generate it randomly. This led to a possible
infinite loop where if r or s were invalid, the loop would continue forever.
This commit fixes this by picking a random value of k in such cases (which are
rare), but this is technically incorrect since it is incompatible with
determinstic k. I may need to rethink how I distinguish between random k and
deterministic k to resolve this issue correctly.