support bip 62 low s; add test vector
I wanted to complete the BitcoinJS test vector which ensures the value of r and
s are computed exactly correclty. However, BitcoinJS was computing signatures
using the BIP 62 low s rule. BIP 62 is designed to reduce transaction
malleability, so signatures should be produced according to this form, and at
some point in the future signatures not conforming to BIP 62 will be regarded
as invalid, which will cure transaction malleability. Since the BitcoinJS test
vector depended on the BIP 62 low s value, I had to add support for BIP 62 low
s to ECDSA to compute the same s value as BitcoinJS. Further, I had to change
the tests for sig2pubkey which were depending on computing the high s value.
Now both low and high s values are tested with sig2pubkey.

Reference:
* https://github.com/bitcoin/bips/blob/master/bip-0062.mediawiki#Low_S_values_in_signatures