remove random element from default signing
In the rare case where the computed value of k is unacceptable, rather than
falling back to a random value, throw an error. This can and should be fixed in
a future update where the full algorithm for deterministic k from RFC 6967

https://tools.ietf.org/html/rfc6979#section-3.2

...also, remove the obsolete signDeterministicK function. Since k is now
deterministic by default, that function is not useful.