make all script chunks the same type
It is slightly confusing and inconsistent that script chunks can either be a
number or an object. I did it that way because BitcoinJS+bitcore did it that
way. However, I find it to be more intuitive if they are all the same type. And
that type is just an object with an opcodenum property, and optionally a
buffer, and optionally also a len, for PUSHDATA ops where the len may or may
not be equivalent to the buf.length (whether this ever occurs in the blockchain
I don't know, but it is a logical possibility).