CheckMinimalPush
When pushing data to the stack, sometimes there is more than one way to do it.
For instance, you can OP_1, which is the simplest way to push 1 to the stack.
However, you could also do OP_PUSHDATA1, followed by 1 (the amount of data),
and then a 1 (the data). That is another way to push 1 to the stack, but
consumes 3 bytes rather than just one. CheckMinimalPush tries to check that a
script is pushing data in the minimal way.

Note that in bitcoind, this function is called CheckMinimalPush, and as of the
time of this writing is located in interpreter.cpp. However, it is more
appropriate for this function to be located in script.js in fullnode. So that
is where the core logic is, and there is a bitcoind-like function that uses the
function from script.js.