v0.1.0
Now that the script interpreter is validated against script_valid and
script_invalid tests, this is a good time to bump up to 0.1.0 territory. From
now on, the versioning we will use wil be compatible with something like semver
( http://semver.org/ ).

Major.Minor.Patch

Or since we're not yet at 1.0.0, we will regard this as:

0.Major.Minor

Major will be iterated when API-breaking changes occur. Otherwise, Minor will
be iterated.