add remaining tx_invalid test
This test was failing and I had suspected there was something wrong with my
sighash code, however it turns out that was not the problem. This test actually
has one script that failes, and the other that succeeds. I had written the test
incorrectly, assuming that all scripts would fail. The fact that one succeeded
counfounded my test. There was nothing actually wrong with my sighash code or
script interpreter code, at least so far as I can tell.