rely on browserify for browser random
I had previously written code that used node's randomBytes in node, and the
browser's crypto.getRandomValues in the browser. This code was originally
written in bitcore, and the copy+pasted into now what is fullnode.

However, blockchain.info's recent security problem has made me rethink the way
to handle random numbers. Here's their issue:

https://www.reddit.com/r/Bitcoin/comments/2onm5r/blockchaininfo_security_disclosure/

My thinking is that I would rather simplify the code and make it easier to
audit. We can see that browserify does the right thing in using
crypto.getRandomValues in the browser:

https://github.com/dominictarr/crypto-browserify/blob/6602891ba0e28b6e0979443e85a73dba8c010bb8/rng.js

browserify's random values uses crypto.getRandomValues when available, and when
not, it throws an error. This is the correct way to handle this problem. Since
we are already thoroughly dependent on browserify, rather than have an
independent implementation of this in fullnode, we'll just adopt browserify's
correct solution.

This also has the benefit that it is easier to ensure full code coverage as
reported by istanbul. random.js now has 100% code coverage, so that we can more
easily see that it is being tested correctly both in node and the browser.

The downside is that we are now more dependent on browserify. I'm not sure how
to handle that in the long run, as I'd rather not write my own build tool to
replace browserify.