sigError -> verifystr
In anticipation of using the same function names in txbuilder, verifystr is a
slightly more general and explanatory name. Every "verify" function in various
classes could have a corresponding "verifystr" function that upon error returns
a string representing that the error was. This same convention will be used in
txbuilder.