One extra loop is required in bad rs computation
A bug reporter has reported that this same logic in bitcore has a problem, see
here:

https://github.com/bitpay/bitcore/pull/884

A re-reading of RFC 6969 indicates that an extra HMAC computation is performed
when badrs > 0. This fixes this possible error, but not extra test cases are
included, because I'm not aware of a test case when badrs > 0.