always trigger second HMAC on badrs or invalid k
When either bad rs is triggered or k is out of bounds, there was a missing HMAC
according to RFC 6979. The second HMAC actually occurs at the start of the
algorithm described in RFC 6979, and has a slightly different layout than the
code here, so the error was hard to find.

The previous commit, which claimed to fix this, was actually incorrect. The
loop was actually only being triggered when badrs > 0 or k is invalid. Thus, it
is not necessary to check for i > 0. It would be helpful to find a test-case
for a signature produced that triggers badrs, and compare that to every
implementation of RFC 6979.