Message -> BSM
The name "Message", which referred to a Bitcoin Signed Message, was too
confusingly named since it implies a "p2p message" even though it has nothing
to do with the p2p messages of bitcoin. Thus it has been renamed to "BSM",
short for "Bitcoin Signed Message", so to be less confusing.