npm run build
Update the browserify dependency, and re-add "npm run build" to run the build.
The previous error which had resulted in a revert was that I had been running a
different version of browserify locally than was being required as a dependency
in fullnode. The 5.x.x. browserify was simply slower. The newer 8.x.x version
is faster.