pass buggy BN test vector
I had a test vector to ensure bn.js is fixed after its recent but, and also
expose BN's static elements which are required for the test to work. This test
vector is from this bn.js commit:

https://github.com/indutny/bn.js/commit/3557d780b07ed0ed301e128f326f83c2226fb679

I am importing it into fullnode to ensure that it stays fixed. See this related
commit in fullnode where I updated the bn.js dependency to fix the error this
caused:

dbeab00365a837ee2a272cf123acde811202afce