add AES+CBC test vector from sjcl
We take the CBC tests from sjcl, which actually use a combination of CBC and
AES, to test our AES+CBC implementation. Note that the sjcl tests were
generated using pkcs#5, and we use pkcs#7, but for our purposes these are
equivalent (pkcs#7 just allows arbitrary block sizes), so the tests pass
unchanged.