test against OpenSSL bug CVE-2014-3570
Bitcoin developer Peter Wuile discovered a bug in OpenSSL in the course
of developing libsecp256k. The OpenSSL security advisory is here:

https://www.openssl.org/news/secadv_20150108.txt

Greg Maxwell has a description of the bug and how it was found here:

https://www.reddit.com/r/Bitcoin/comments/2rrxq7/on_why_010s_release_notes_say_we_have_reason_to/
https://www.reddit.com/r/programming/comments/2rrc64/openssl_security_advisory_new_openssl_releases/

The OpenSSL fix is here:

https://github.com/openssl/openssl/commit/a7a44ba55cb4f884c6bc9ceac90072dea38e66d0

This fullnode test uses the same big numbers that were problematic in OpenSSL.
The check is to ensure that squaring a number and multiplying it by itself
result in the same number. As an additional precaution, we check this
multiplication normally as well as mod(p).