var -> let
"let" is a strictly safer version of "var", available in ES6. I've updated the
entire library to use "let" instead of "var". Along the way, I (necessarily)
fixed some bad practices that were valid with "var", but invalid with "let",
such as declaring a variable multiple times, or declaring it in the wrong block
scope. Additionally, I found one bug, which was with compressed public keys
when finding the public key from a compact signature. The check on i was being
performed before i was known; valid with "var", and invalid with "let".

Note that this commit means that the source code for fullnode does not run on
vanilla node any longer; it only works on io.js or node harmony. To run the
code in vanilla node, you can build it with an ES6 transpiler, such as es6ify,
which is what I'm using to build it for the browser.