distinguish between Default, Mainnet, Testnet
One of my goals is to make fullnode configurable at every level. Right now some
classes are configurable by injecting dependencies into them (BIP32, Address,
Privkey). You can also configure it at an app level by choosing which classes
to include. I am now building Mainnet and Testnet configurations by default
which, right now, simply have different constants for a few classes. Eventually
there will be different databases and network systems.

There is a difference between "Default", "Mainnet" and "Testnet". Mainnet is
mainnet settings, and Testnet is testnet settings, of course, and they are not
intended to be overridden. Default, however, is a copy of the Mainnet
configuration, but is intended to be overridden. In the future, when the
database and network are also included in the configuration, you will want to
override the defaults with ports and such. Thus, although right now Default and
Mainnet will be the same, Default will end up being something that a user of
fullnode will actually edit, whereas Mainnet will not.