expose injected dependencies; add extend
Rather than place injected dependencies on the constructor, which was a bit
weird (why should Address.Base58Check exist?), I'm now placing them on an
"injected" variable that is on the constructor. The benefit is that, first of
all, it looks nicer (Address.injected.Base58Check is more obviously an injected
dependency rather than a property of Address), and also the object is the exact
same as the one that was passed in. This means if you need to recreate a class
for some reason, you can, using an API like this:

let Address2 = Address.inject(Address.injected);
assert(Address2 === Address); // true

I've also added an "extend" utility, like jQuery's extend or other libraries. I
would have preferred to use ES6's Object.assign(), but unfortunately that is
not yet available by default in io.js. Oh well.