bw.concat() -> bw.toBuffer()
The Buffer Writer is obviously all about writing buffers. You want to combine a
bunch of things together and turn it into a buffer. Therefore the .toBuffer()
method makes perfect sense, and is compatible with the rest of the library.
.concat() is a technical term that simply comes from the fact that the way this
is implemented it to use the Buffer.concat() method. However, for ease-of-use
and consistency, I think sticking with .toBuffer() exclusively is the best
approach.