typeof x === 'undefined' -> x === undefined
I've been using the typeof operator to check to see if a variable is undefined
due to paranoia about javascript's type coercion, but as far as I can tell, it
is completely equivalent to use a triple equals to check if it is undefined.
The only difference is if the variable has never been declared, but I never
actually need to check that in fullnode.