use generators for interp
It is useful to be able to step through a script that is interpreting,
and generators + iterators are perfect for this. The eval function is
now a generator, so you can iterate through it. The verify function has
been replaced with results, which is the equivalent of bitcoin core's
EvalScript method. There is a new verify function, with the same
interface as the old, that iterates through results.