fromBuffers, expect
I'm beginning to implement the bitcoin p2p protocol, and the most
important element the interface to most classes that is missing is a way
to stream in data. Rather than accept an actual stream, e.g. fromStream,
or fromBufferStream, I've implemented a tool designed to be used with a
promises/generators pattern. fromBuffers is a generator that creates an
iterator that lets you pass in buffers of any size to slowly build an
object by passing in the buffers via .next(buf). I have implemented it
on msg for the p2p protocol, but will ultimately implement it on all
other "struct" classes. I have also implemented a convenience method
.expect which is a generator returning an iterator that lets you pass in
buffers until you reach the expected length, and then it returns the
remainder.