break apart worker-node and worker-browser
Although I had only just recently merged them, I realized there is a
significant problem in using the same file for both workers - browserify
was adding in a bunch of irrelevant stuff to the build making the worker
bundle enormous. I tried using browserify config and some other
obscuration techniques to remove the extra stuff, but that was really
cumbersome and didn't completely work. The best way to solve this seems
to me to keep the two files completely separate. The top several lines
are now exactly the same, and the only differences are at the bottom,
where the different management code lives. I may be able to improve this
situation in the future by using a better build tool to concatenate the
part that is the exact same. But for now, it is two separate files.