expose Connection, Server, Work, Workers
Exposing these in index.js has revealed a minor bug with web workers,
which is that the "Worker" global object actually does not exist in a
web worker. Instead, you must use "self", which is the web worker's
global object, and then access Worker in self, i.e. self.Worker. At
least that's how it is in Chrome, Firefox didn't seem to have an issue
with the Worker object in a web worker.